+++ tests/array_read.cpp
@@ -39,6 +39,6 @@ int main() {
   chunk_bytes.reset();
 
   // Cleanup
-  zarrsDestroyArray(array);
-  zarrsDestroyStorage(storage);
+  zarrs_assert(zarrsDestroyArray(array));
+  zarrs_assert(zarrsDestroyStorage(storage));
 }